pythonBackend/requirements.txt
fastapi
uvicorn
strawberry-graphql
httpie
black
isort
torch
prisma
redis
pprintpp
pylatex[matrices]
jinja2
pydantic
python-ulid[pydantic]
requests
langchain
langchain_ollama
typesense
pika
huggingface_hub
python-multipart
transformers
accelerate
flash-attn
ffmpeg
python-dotenv
pyjwt
ffmpeg-python
pillow
ollama
pyzbar 
opencv-python
biip
pycdlib
pydash

